Proper setup, path configuration, and cleanup
Updated readme with installation guide and more recent description. Did
a lot of path configuration to base everything from the top level
gridbots package, and add this to the path if needed in run.py. Added
docstrings to more classes. Updated requirements.txt. Overall cleanup to
reflect the state of things.

Signed-off-by: Hayk Martirosyan <hayk.mart@gmail.com>