Bugfix: Stuck when bot starts on goal
Previously, if a bot was assigned a goal on its current position, it
would go to an at_home state and mark its goal as None. Then, the
at_goal function would return False, making it look like it is not done
yet. Now, planning module checks if at_goal or has_goal.

Signed-off-by: Hayk Martirosyan <hayk.mart@gmail.com>