Improved script running
Added more script files. Allowed a couple of common typos in scripts,
to not have to modify the scripts individually and they can stay the
same as they were sent to me. Improved bot handling of multiple
points to move. Removed grid points around indexing area so that
indexing looks right (is pointless in simulation, but must match up
with the real robot behavior). Made the routine generator run
automatically whenever we run a simulation.