Split up controller from simulation, optimize for speed
Optimization study in README. Seperate out controller logic into its own
class, so that the controller outputs zone commands and the bots just
follow them. The controller also decides when the simulation is over.
Created a SingleRoutineController to just test tlap. Eventually, a full
controller will parse the structure and run routines to build it.