Filled mesh and autogenerated build!!
The simulation now specifies a structure in the form of a .ply file, and
generates its build script before running it. The .ply file is in
specs/structures and is expected to have only orthogonal edges, and all
edges of the same length (which get scaled to the lattice size of the
rods when being processed).